Remove memory channel storage
